refactor(goods): 상품 상세 커스텀 훅 및 쿼리 옵션을 FSD 구조에 맞춰 model 및 lib로 분리
